feat: Load midi elements list dynamically
