+++ packages/midi-bricks/src/reducers/index.js
@@ -11,7 +11,7 @@ export const PAGE_TYPES = {
   HOME_MODE: 'HOME_MODE',
   GLOBAL_MODE: 'GLOBAL_MODE',
   MIDI_DRIVER_MODE: 'MIDI_DRIVER_MODE',
-  VIEW_SETTINGS_MODE: 'VIEW_SETTINGS_MODE'
+  // VIEW_SETTINGS_MODE: 'VIEW_SETTINGS_MODE'
 }
 
 const initState = {
